Inline TransformaionNodeFactory
